+++ CHANGELOG.md
@@ -18,6 +18,12 @@
 * Incorrect server name verification using Postgres in `verify-ca` mode 
 * Tx-sender exits sending Tessera private transaction with invalid 'from'
 
+## v21.1.11 (2021-11-23)
+
+### 🛠 Bug fixes
+* Tx-sender exits updating jobs already in final status
+* Tx-sender does not send message on `tx-recover` topic when there are persistent connectivity issues with RPC nodes
+
 ## v21.1.10 (2021-10-28)
 
 ### 🛠 Bug fixes